Allow piping lists of people
