+++ lib/pco_api/actions.ex
@@ -6,7 +6,7 @@ defmodule PcoApi.Actions do
 
       def request(:get, url, params) do
         case get(url, [], params: params, hackney: [basic_auth: {PcoApi.key, PcoApi.secret}]) do
-          {:ok, %HTTPoison.Response{status_code: 200, body: body}} ->
+          {:ok, %HTTPoison.Response{status_code: code, body: body}} when (code in 200..299) ->
             body["data"]
           {:ok, %HTTPoison.Response{body: body}} ->
             %{"errors" => [%{"detail" => detail, "title" => title}]} = body
@@ -14,6 +14,26 @@ defmodule PcoApi.Actions do
           {:error, _} ->
             raise "PcoApi.People error"
         end
+      end
+
+      def get, do: get("")
+      def get(id) when is_integer(id), do: get(Integer.to_string(id))
+      def get(url) when is_binary(url), do: request(:get, url, []) |> new
+      def get(params) when is_list(params), do: get(params, "")
+      def get(params, url) when is_list(params), do: request(:get, url, params) |> new
+
+      def get_list([%PcoApi.Record{} | rest] = records), do: get_one(records)
+      def get_one([]), do: []
+      def get_one(%PcoApi.Record{} = record) do
+        get record.links["self"]
+      end
+      def get_one([%PcoApi.Record{} = first | rest]) do
+        [get_one(first) | get_one(rest)]
+      end
+
+      def new(results) when is_list(results), do: results |> Enum.map(&(&1 |> new))
+      def new(%{"id" => id, "links" => links, "attributes" => attrs, "type" => type}) do
+        %PcoApi.Record{id: id, links: links, attributes: attrs, type: type}
       end
 
       # if this is a full URL, don't add the endpoint. This allows using direct links.
@@ -31,19 +51,6 @@ defmodule PcoApi.Actions do
   defmacro endpoint(url) do
     quote do
       def unquote(:api_endpoint)(), do: unquote(url)
-    end
-  end
-
-  # TODO: this probably doesn't belong here
-  defmacro linked_association(name) do
-    quote do
-      def unquote(:"#{name}")(record) do
-        # TODO: error handling when the association link doesn't exist
-        request(:get, record.links[Atom.to_string(unquote(name))], [])
-        |> Enum.map(fn(%{"attributes" => attrs, "id" => id, "links" => links, "type" => type}) ->
-          %PcoApi.Record{attributes: attrs, id: id, links: links, type: type}
-        end)
-      end
     end
   end
 end
